GDrive mocks are specific to sonny source
